Display wiki categories for a submitted Knewton category path
The method tries to come up with the most relevant wiki
category paths to explore for the given Knewton (or whatever)
category path.

The algorithm it uses is as follows:
1. For a given query, e.g. Algebra & Algebra 2|Matrix Algebra|Properties of Matrices,
get the first certain number (configured) of wiki search results.
2. Gather the publicly displayed categories these articles belong to,
and compute the number of words in these sets that also occur in the search query.
3. For the most relevant result (most common words), display the found
wiki category list, and the article determined to be the most relevant
from the given retrieved sample.

For instance, for the above mentioned query the result is:
Matrix unit: Matrix theory|Semigroup theory

Usage:
Simply type: python get_wiki_categories.py, the else is clear from the prompt.

Todo: I'll add config options to the interface, to allow to play with it
in a more flexible way.