Refining the algorithm, adding batch mode, and documentation
The method now allows to specify a file with one taxonomy path/concept
per line, and the name of the output file.

It computes the lists of relevant topics for each taxon paths, with some
additional information like which page was used to get those.

Also added the file of retail taxonomy taxons, parsed from:
https://beta.knewton.com/goals/assign, that can be used for testing
the algorithm.