Improving the category/relevant link script
The output is now in two files (if file output is desired):
1. _categories.txt file is listing the relevant categories per taxonomy path
2. _relevant_links.txt is listing relevant wikipedia page titles per
taxonomy path

I also added data responses I got for the retail math taxonomy forest.
(For each taxonomy path, 10 search results were examined, 3 of the best
based on categorization chosen for the best category matches, and the links
they shared for the best link suggestions)

TODO: Enhance the relevant link file with urls. Remove comma as a delimiter.